Update create-compliance-policy.md
I tried to find in several MS docs for info on how is the Compliance State of a specific device updated. There is reference on this doc https://docs.microsoft.com/en-us/mem/intune/configuration/device-profile-troubleshoot#how-long-does-it-take-for-devices-to-get-a-policy-profile-or-app-after-they-are-assigned to a 8 hour period to automatically reevaluate compliance state but I haven't found anywhere information about being able to force the update of the compliance state.

In the doc Create Compliance Policy, there is reference to Refresh cycle times. Here there is info about policy update but not specifically to compliance state update. As it is, my interpretation of the document is that I can force the synchronization but it will only update the settings of the assigned policy(ies) if there are any assigned. Compliance state is not reevaluated.

With the change i propose, the document gets changed to have that detail as to make absolutely clear that there is no need to wait for the 8 hour refresh cycle.